Update geodata.m
Tried to reorganize so it would take into account the actual extents of the dem read.
Also I reduced memory check to size of the single precision and changed stride to be computed based on the stride necessary to satsify the 4 GB limit. Keep comments for your reference to try it out on your end.